Rename types -> module
