--- src/engine/types/types.game.ts
@@ -1,7 +1,0 @@
-import { GameState } from "../GameWorker";
-
-export interface GameThreadModule<ModuleState> {
-  create(): ModuleState;
-  init: (state: GameState) => Promise<void>;
-  dispose?: (state: GameState) => void;
-}
