--- src/engine/types/types.render.ts
@@ -1,7 +1,0 @@
-import { RenderThreadState } from "../RenderWorker";
-
-export interface RenderThreadModule<ModuleState> {
-  create(): ModuleState;
-  init: (state: RenderThreadState) => Promise<void>;
-  dispose?: (state: RenderThreadState) => void;
-}
